OVRO-LWA check.

+++ update_casa_data.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-mkdir -p /home/ubuntu/anaconda3/envs/casa/lib/casa/data/
-cd /home/ubuntu/anaconda3/envs/casa/lib/casa/data/
-rsync -avz rsync://casa-rsync.nrao.edu/casa-data .
+if [[ -e /home/ubuntu/anaconda3/envs/casa ]]; then
+  mkdir -p /home/ubuntu/anaconda3/envs/casa/lib/casa/data/
+  cd /home/ubuntu/anaconda3/envs/casa/lib/casa/data/
+  rsync -avz rsync://casa-rsync.nrao.edu/casa-data .
+fi
